+++ src/exceptions/FaceIndexError.ts
@@ -1,0 +1,17 @@
+/**
+ * Face index error.
+ */
+export default class extends Error {
+  /**
+   * @type {string}
+   */
+  public name: string;
+
+  /**
+   * @param {string} message Error message.
+   */
+  constructor(message: string) {
+    super(message);
+    this.name = 'FaceIndexError';
+  }
+}