+++ src/exceptions/MissingFaceInImageError.ts
@@ -1,0 +1,17 @@
+/**
+ * Missing face in image error.
+ */
+export default class extends Error {
+  /**
+   * @type {string}
+   */
+  public name: string;
+
+  /**
+   * @param {string} message Error message.
+   */
+  constructor(message: string = 'There is no face in the image') {
+    super(message);
+    this.name = 'MissingFaceInImageError';
+  }
+}