+++ types/exceptions/index.d.ts
@@ -1,2 +1,6 @@
-import ApiError from './ApiError';
-export { ApiError };
+import CollectionCreationError from './CollectionCreationError';
+import CollectionDeletionError from './CollectionDeletionError';
+import FaceIndexError from './FaceIndexError';
+import MissingFaceInImageError from './MissingFaceInImageError';
+import MultipleFacesInImageError from './MultipleFacesInImageError';
+export { CollectionCreationError, CollectionDeletionError, FaceIndexError, MissingFaceInImageError, MultipleFacesInImageError, };